+++ artist_type/types.go
@@ -1,0 +1,8 @@
+package artist_type
+
+type ArtistType string
+
+const (
+	ARTIST      ArtistType = "ARTIST"
+	CONTRIBUTOR ArtistType = "CONTRIBUTOR"
+)