+++ encryption_type/types.go
@@ -1,0 +1,7 @@
+package encryption_type
+
+type EncryptionType string
+
+const (
+	None EncryptionType = "NONE"
+)