+++ video_quality/types.go
@@ -1,0 +1,11 @@
+package video_quality
+
+type VideoQuality int
+
+const (
+	P240  VideoQuality = 240
+	P360  VideoQuality = 360
+	P480  VideoQuality = 480
+	P720  VideoQuality = 720
+	P1080 VideoQuality = 1080
+)