+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [2.1.0] - 2025-02-03
+
+### Added
+
+- Track download method
+- Track download to file method
+- Video download method
+- Video download to file method
+
+### Changed
+
+- Remove panics from request sender
+
 ## [2.0.1] - 2025-02-03
 
 ### Changed